Add new ES6 test file
The previously test file did not use deterministically generated
test inputs, which made it impossible for implementations to verify
correctness without fetching the 2GiB test file.

The new test file is deterministically generated, allowing implementations
to locally verify correctness without any network bandwidth.

As a minor note, we compress the file using GZIP instead of ZIP
so that the test file can be decompressed in a streaming manner
as it is being transferred over the network.